feat(services): do not suppress config update on installation
It was meant to leave the configuration untouched on boot,
especially useful when multiple motion services are enabled.
Unfortunately this has the effect that enabling a service that is
not enabled on boot will not update the config either, so
the user has to enable a service twice to see an effect.